+++ doc_source/_ext/test_cw_resources.py
@@ -28,6 +28,8 @@ class TestCloudWandererResourcesDirective(unittest.TestCase):
     * :class:`RestApis<apigateway.rest_api>`
 * :doc:`Lambda <resource_properties/lambda>`
     * :class:`Functions<lambda.function>`
+    * :class:`Layers<lambda.layer>`
+         * :class:`Layer Versions<lambda.layer.layer_version>`
 * :doc:`Secrets Manager <resource_properties/secretsmanager>`
     * :class:`Secrets<secretsmanager.secret>`
 """