+++ tests/integration/custom_resources/lambda/test_layers.py
@@ -19,14 +19,20 @@ class TestLambdaLayers(NoMotoMock, unittest.TestCase):
         },
     }
 
+    layer_version_payload = {
+        "LayerVersionArn": "arn:aws:lambda:eu-west-1:123456789012:layer:test-layer:1",
+        "Version": 1,
+        "Description": "This is a test layer!",
+        "CreatedDate": "2020-10-17T13:18:00.303+0000",
+        "CompatibleRuntimes": ["nodejs10.x"],
+    }
+
     mock = {
         "lambda": {
-            "get_layer.return_value": {"Configuration": layer_payload},
-            "get_paginator.return_value.paginate.return_value": [
-                {
-                    "Layers": [layer_payload],
-                }
-            ],
+            "list_layers.return_value": {
+                "Layers": [layer_payload],
+            },
+            "list_layer_versions.return_value": {"LayerVersions": [layer_version_payload]},
         }
     }
 
@@ -39,6 +45,6 @@ class TestLambdaLayers(NoMotoMock, unittest.TestCase):
     multiple_resource_scenarios = [
         MultipleResourceScenario(
             arguments=CloudWandererCalls(regions=["eu-west-1"], service_names=["lambda"], resource_types=["layer"]),
-            expected_results=[layer_payload],
+            expected_results=[layer_payload, layer_version_payload],
         )
     ]